Revert "add dependencies, failing on direct import of fs"
This reverts commit f012e9511869acdc17e28a34ccc43f4e98c4610e.
