+++ package.json
@@ -14,7 +14,7 @@
     "format": "prettier --write src",
     "format:check": "prettier -c src",
     "lint": "eslint .",
-    "start": "webpack serve --config ./webpack.serve.config.ts",
+    "start": "node ./lib/Bundler/source-map-compare",
     "test": "jest"
   },
   "bin": {
@@ -27,32 +27,20 @@
     "@caporal/core": "^2.0.2",
     "@fluentui/react-components": "^9.48.0",
     "@fluentui/react-icons": "^2.0.235",
-    "assert": "^2.1.0",
-    "browserify-zlib": "^0.2.0",
-    "buffer": "^6.0.3",
-    "css-loader": "^7.1.1",
     "glob": "^10.3.12",
     "html-webpack-plugin": "^5.6.0",
-    "os-browserify": "^0.3.0",
     "path-browserify": "^1.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "rimraf": "^5.0.5",
     "source-map": "^0.7.4",
     "source-map-explorer": "2.5.3",
-    "stream-browserify": "^3.0.0",
-    "style-loader": "^4.0.0",
     "tslib": "^2.6.2",
-    "url": "^0.11.3",
-    "use-debounce": "^10.0.0",
-    "util": "^0.12.5",
-    "webpack": "^5.91.0",
-    "webpack-merge": "^5.10.0"
+    "webpack": "^5.91.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.0.0",
     "@microsoft/eslint-plugin-sdl": "^0.2.2",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.11",
     "@types/jest": "^29.5.12",
     "@types/react": "^18.2.79",
     "@types/react-dom": "^18.2.25",
@@ -65,19 +53,12 @@
     "eslint-plugin-jest": "^28.2.0",
     "eslint-plugin-jsx-a11y": "^6.8.0",
     "eslint-plugin-prettier": "^5.1.3",
-    "fork-ts-checker-notifier-webpack-plugin": "^9.0.0",
-    "fork-ts-checker-webpack-plugin": "^9.0.2",
     "jest": "^29.7.0",
     "prettier": "^3.2.5",
-    "react-refresh": "^0.14.0",
-    "react-refresh-typescript": "^2.0.9",
     "ts-jest": "^29.1.2",
     "ts-loader": "^9.5.1",
-    "ts-node": "^10.9.2",
     "typescript": "^5.4.3",
-    "typescript-eslint": "^7.7.0",
-    "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^5.0.4"
+    "typescript-eslint": "^7.7.0"
   },
   "resolutions": {
     "**/lodash": ">=4.17.21",
