--- public/index.html
@@ -1,33 +1,0 @@
-<!DOCTYPE html>
-
-<html lang="en-us">
-
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <title>Bundle Size Viewer</title>
-  <style>
-    html,
-    body {
-      margin: 0;
-      height: 100%;
-      background-color: var(--colorNeutralBackground1, #fff);
-    }
-
-    @media (prefers-color-scheme: dark) {
-      body {
-        background-color: var(--colorNeutralBackground1, rgb(41, 41, 41));
-      }
-    }
-
-    #root {
-      height: 100%;
-    }
-  </style>
-</head>
-
-<body>
-  <div id="root"></div>
-</body>
-
-</html>
