--- src/Bundler/browser/process.ts
@@ -1,1 +1,0 @@
-export const platform = 'linux';
