+++ src/Bundler/inspector-browser.ts
