--- src/FileSystem/WebFS/index.ts
@@ -1,1 +1,0 @@
-export * from './WebFileSystem';
