--- src/View/OpenDialog.styles.ts
@@ -1,18 +1,0 @@
-import { makeStyles, shorthands, tokens } from '@fluentui/react-components';
-
-export const useStyles = makeStyles({
-  modeButton: {
-    ...shorthands.flex(1)
-  },
-  dialogBody: {
-    display: 'flex',
-    flexDirection: 'column',
-    ...shorthands.gap(tokens.spacingVerticalM)
-  },
-  modeSelection: {
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
-    ...shorthands.gap(tokens.spacingVerticalXS)
-  }
-});
