--- src/View/Root.styles.ts
@@ -1,7 +1,0 @@
-import { makeStyles } from '@fluentui/react-components';
-
-export const useStyles = makeStyles({
-  root: {
-    height: '100%'
-  }
-});
